Some renaming for consistency
